requirements.txt
streamlit>=1.28.0
openai-whisper
streamlit-audiorecorder
